+++ Juyoung/vue-project/db.json
@@ -81,7 +81,7 @@
       "id": 21
     },
     {
-      "subject": "new todo223",
+      "subject": "new todo22",
       "completed": false,
       "id": 22
     }
