+++ Juyoung/vue-project/src/pages/todos/create/index.vue
@@ -1,10 +1,16 @@
 <template>
-  create page
+  <div>
+    <h1>Create Todo</h1>
+    <TodoForm />
+  </div>
 </template>
 
 <script>
+import TodoForm from '@/components/TodoForm.vue'
 export default {
-
+  components: { 
+    TodoForm 
+  }
 }
 </script>
