+++ Juyoung/vue-project/src/pages/todos/index.vue
@@ -17,9 +17,6 @@
       @keyup.enter="searchTodo"
     >
     <hr />
-    <TodoSimpleForm @add-todo="addTodo"/>
-    <div style="color: red">{{ error }}</div>
-
     <div v-if="!todos.length">
       There is nothing to display
     </div>
@@ -58,7 +55,6 @@
 
 <script>
 import { ref, computed, watch } from 'vue';
-import TodoSimpleForm from '@/components/TodoSimpleForm.vue';
 import TodoList from '@/components/TodoList.vue';
 import axios from 'axios';
 import Toast from '@/components/Toast.vue';
@@ -67,7 +63,6 @@ import { useRouter } from 'vue-router';
 
 export default {
   components: {
-    TodoSimpleForm,
     TodoList,
     Toast
   },
